+++ app/src/main/java/org/chromium/blink/mojom/AppCacheErrorReason.java
@@ -32,8 +32,8 @@ public final class AppCacheErrorReason {
 
     public static final int APPCACHE_UNKNOWN_ERROR = APPCACHE_POLICY_ERROR + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (7);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 7;
 
     private static final boolean IS_EXTENSIBLE = false;
