+++ app/src/main/java/org/chromium/blink/mojom/AppCacheStatus.java
@@ -28,8 +28,8 @@ public final class AppCacheStatus {
 
     public static final int APPCACHE_STATUS_OBSOLETE = (int) (5L);
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (5);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 5;
 
     private static final boolean IS_EXTENSIBLE = false;
