+++ app/src/main/java/org/chromium/blink/mojom/AttestationConveyancePreference.java
@@ -24,8 +24,8 @@ public final class AttestationConveyancePreference {
 
     public static final int ENTERPRISE = DIRECT + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (3);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 3;
 
     private static final boolean IS_EXTENSIBLE = false;
