+++ app/src/main/java/org/chromium/blink/mojom/BackgroundFetchError.java
@@ -34,8 +34,8 @@ public final class BackgroundFetchError {
 
     public static final int REGISTRATION_LIMIT_EXCEEDED = PERMISSION_DENIED + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (8);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 8;
 
     private static final boolean IS_EXTENSIBLE = false;
