+++ app/src/main/java/org/chromium/blink/mojom/ControllerServiceWorkerPurpose.java
@@ -18,8 +18,8 @@ public final class ControllerServiceWorkerPurpose {
 
     public static final int FETCH_SUB_RESOURCE = 0;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (0);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 0;
 
     private static final boolean IS_EXTENSIBLE = false;
