+++ app/src/main/java/org/chromium/blink/mojom/FeaturePolicyFeature.java
@@ -80,8 +80,8 @@ public final class FeaturePolicyFeature {
 
     public static final int FONT_DISPLAY = (int) (32L);
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (32);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 32;
 
     private static final boolean IS_EXTENSIBLE = false;
