+++ app/src/main/java/org/chromium/blink/mojom/FetchCacheMode.java
@@ -32,8 +32,8 @@ public final class FetchCacheMode {
 
     public static final int UNSPECIFIED_FORCE_CACHE_MISS = UNSPECIFIED_ONLY_IF_CACHED_STRICT + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (7);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 7;
 
     private static final boolean IS_EXTENSIBLE = false;
