+++ app/src/main/java/org/chromium/blink/mojom/ManifestShareTarget.java
@@ -27,8 +27,8 @@ public final class ManifestShareTarget extends org.chromium.mojo.bindings.Struct
 
         public static final int POST = GET + 1;
 
-        public static final int MIN_VALUE = (int) (0);
-        public static final int MAX_VALUE = (int) (1);
+        public static final int MIN_VALUE = 0;
+        public static final int MAX_VALUE = 1;
 
         private static final boolean IS_EXTENSIBLE = false;
 
@@ -59,8 +59,8 @@ public final class ManifestShareTarget extends org.chromium.mojo.bindings.Struct
 
         public static final int MULTIPART = APPLICATION + 1;
 
-        public static final int MIN_VALUE = (int) (0);
-        public static final int MAX_VALUE = (int) (1);
+        public static final int MIN_VALUE = 0;
+        public static final int MAX_VALUE = 1;
 
         private static final boolean IS_EXTENSIBLE = false;
