+++ app/src/main/java/org/chromium/blink/mojom/QuotaStatusCode.java
@@ -28,8 +28,8 @@ public final class QuotaStatusCode {
 
     public static final int UNKNOWN = (int) (-1L);
 
-    public static final int MIN_VALUE = (int) (-1);
-    public static final int MAX_VALUE = (int) (20);
+    public static final int MIN_VALUE = -1;
+    public static final int MAX_VALUE = 20;
 
     private static final boolean IS_EXTENSIBLE = false;
