+++ app/src/main/java/org/chromium/blink/mojom/ServiceWorkerClientType.java
@@ -22,8 +22,8 @@ public final class ServiceWorkerClientType {
 
     public static final int ALL = SHARED_WORKER + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (2);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 2;
 
     private static final boolean IS_EXTENSIBLE = false;
