+++ app/src/main/java/org/chromium/blink/mojom/ServiceWorkerErrorType.java
@@ -44,8 +44,8 @@ public final class ServiceWorkerErrorType {
 
     public static final int UNKNOWN = TYPE + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (13);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 13;
 
     private static final boolean IS_EXTENSIBLE = false;
