+++ app/src/main/java/org/chromium/blink/mojom/ServiceWorkerEventStatus.java
@@ -22,10 +22,10 @@ public final class ServiceWorkerEventStatus {
 
     public static final int ABORTED = REJECTED + 1;
 
-    public static final int MAX = (int) (ServiceWorkerEventStatus.ABORTED);
+    public static final int MAX = ServiceWorkerEventStatus.ABORTED;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (2);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 2;
 
     private static final boolean IS_EXTENSIBLE = false;
