+++ app/src/main/java/org/chromium/blink/mojom/WebFeature.java
@@ -4264,8 +4264,8 @@ public final class WebFeature {
 
     public static final int NUMBER_OF_FEATURES = MEDIA_SESSION_SKIP_AD + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (2784);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 2784;
 
     private static final boolean IS_EXTENSIBLE = false;
