+++ app/src/main/java/org/chromium/device/mojom/BiometricType.java
@@ -20,8 +20,8 @@ public final class BiometricType {
 
     public static final int FINGERPRINT = (int) (1L);
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (1);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 1;
 
     private static final boolean IS_EXTENSIBLE = false;
