+++ app/src/main/java/org/chromium/device/mojom/NfcErrorType.java
@@ -34,8 +34,8 @@ public final class NfcErrorType {
 
     public static final int IO_ERROR = CANNOT_CANCEL + 1;
 
-    public static final int MIN_VALUE = (int) (0);
-    public static final int MAX_VALUE = (int) (8);
+    public static final int MIN_VALUE = 0;
+    public static final int MAX_VALUE = 8;
 
     private static final boolean IS_EXTENSIBLE = false;
